[refact]: chat 폴더구조 변경
