+++ src/components/Chat/ChatRoomsEdit/ChatRoomsEdit.module.scss
@@ -4,7 +4,6 @@
   display: flex;
   flex-direction: column;
   justify-content: space-between;
-  height: 100dvh;
 
   > .ChatRoomsEditContainer {
     > .ChatRoomsEditHeader {
