+++ src/pages/Chat/ChatRooms/ChatRoomsPage.tsx
@@ -1,0 +1,7 @@
+import ChatRooms from '../../../components/Chat/ChatRooms/ChatRooms';
+
+const ChatRoomsPage = () => {
+  return <ChatRooms />;
+};
+
+export default ChatRoomsPage;
